% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_model.r
\name{actET}
\alias{actET}
\title{compute actual evapotranspiration}
\usage{
actET(soil_state, soil_param, meteo)
}
\arguments{
\item{soil_state}{is a data frame with soil water content 'theta', 'theta_previous' and excess 'runoff'.}

\item{soil_param}{is a data frame with 'ra', 'rs' and soil parameters as in function _soil_model_.}
}
\value{
new_soil_state
}
\description{
compute actual evapotranspiration
}
