% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{computeEffluent}
\alias{computeEffluent}
\title{compute basin's effluent from all surface processes}
\usage{
computeEffluent(runoff, pipe, structure, subbasin)
}
\arguments{
\item{runoff}{includes volume, runoff_in and runoff_out in subbasin}

\item{pipe}{includes volume, qin and qout in pipe}

\item{structure}{includes volume, qin and qout in structure}
}
\value{
subbasin
}
\description{
compute basin's effluent from all surface processes
}
