% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{gatherAffluentStrahler}
\alias{gatherAffluentStrahler}
\title{gather upstream affluent, only one strahler order can be given, should loop through  strahler in main loop}
\usage{
gatherAffluentStrahler(network.subset, subbasin)
}
\value{
subbasin
}
\description{
gather upstream affluent, only one strahler order can be given, should loop through  strahler in main loop
}
