% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{loop_runoff}
\alias{loop_runoff}
\title{main loop runoff only}
\usage{
loop_runoff(subbasin_initial, I0, strahler)
}
\value{
list of runoff, pipe and structure tbl
}
\description{
main loop runoff only
}
