% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{lossModel}
\alias{lossModel}
\title{compute effective runoff with loss model}
\usage{
lossModel(subbasin)
}
\arguments{
\item{subbasin}{is the global dataframe including rainfall intensity}
}
\value{
runoff
}
\description{
compute effective runoff with loss model
}
