% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_model.r
\name{lossModel_green_ampt}
\alias{lossModel_green_ampt}
\title{compute effective runoff with loss model}
\usage{
lossModel_green_ampt(subbasin)
}
\arguments{
\item{subbasin}{is the global dataframe including rainfall intensity}
}
\value{
runoff
}
\description{
compute effective runoff with loss model
}
