% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{makeSummary}
\alias{makeSummary}
\title{summary catchment}
\usage{
makeSummary()
}
\arguments{
\item{swc}{soil water content}

\item{et}{evapotranspiration}

\item{runoff}{surface runoff}

\item{qaffl}{outlet discharge before structure}

\item{qeffl}{outlet discharge after structure}
}
\description{
summary catchment
}
