% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{routeStructure}
\alias{routeStructure}
\title{route through structure}
\usage{
routeStructure(subbasin)
}
\value{
`structure`, a data frame with columns, `name`, `V`, `Vvirtual`, `Qoverflow`, `Qin`, `Qout`
}
\description{
route through structure
}
