% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_model.r
\name{soil_model}
\alias{soil_model}
\title{compute effective runoff with green-ampt. Needs testing.}
\usage{
soil_model(soil_state, soil_param, meteo)
}
\arguments{
\item{soil_state}{is a data frame with soil water content 'theta', 'theta_previous' and excess 'runoff'.}

\item{soil_param}{is a data frame with soil properties like 'porosity', soil_depth', 'phi' (matric suction depth), 'K' (saturated conductivity), 'dt' (desired time step for green-ampt computation) and 'model_step' (the time step of the model where green ampt is being applied to)}

\item{meteo}{is a data frame with 'i_t' (rainfall intensity in mm/h),}
}
\value{
soil_state
}
\description{
compute effective runoff with green-ampt. Needs testing.
}
