% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{updateSubbasinAfterLossModel}
\alias{updateSubbasinAfterLossModel}
\title{update subbasin dataframe after use of lossModel}
\usage{
updateSubbasinAfterLossModel(subbasin, runoff)
}
\arguments{
\item{runoff}{from lossModel}
}
\value{
subbasin.updated
}
\description{
update subbasin dataframe after use of lossModel
}
