% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{updateSubbasinAfterPipe}
\alias{updateSubbasinAfterPipe}
\title{update subbasin after pipe routing}
\usage{
updateSubbasinAfterPipe(subbasin, pipe)
}
\arguments{
\item{pipe}{as given by routePipe}
}
\value{
subbasin.updated
}
\description{
update subbasin after pipe routing
}
