% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{updateSubbasinAfterRunoff}
\alias{updateSubbasinAfterRunoff}
\title{update subbasin dataframe after runoff was routed}
\usage{
updateSubbasinAfterRunoff(subbasin, runoff)
}
\arguments{
\item{runoff}{as given by routeRunoff}
}
\value{
subbasin.updated
}
\description{
update subbasin dataframe after runoff was routed
}
