% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{updateSubbasinAfterStructure}
\alias{updateSubbasinAfterStructure}
\title{update subbasin after structure}
\usage{
updateSubbasinAfterStructure(subbasin, structure)
}
\arguments{
\item{structure}{as given by routeStructure}
}
\value{
subbasin.updated
}
\description{
update subbasin after structure
}
