% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_states.r
\name{computeEffluent}
\alias{computeEffluent}
\title{compute basin's effluent from all surface processes}
\usage{
computeEffluent(subbasin)
}
\arguments{
\item{subbasin}{is the subbasin state after structure update}
}
\value{
subbasin.effl
}
\description{
compute basin's effluent from all surface processes
}
